Revert long to int in apdu_interface_logic_channel_open

+++ driver/apdu/at.c
@@ -162,7 +162,7 @@ exit:
     return fret;
 }
 
-static long apdu_interface_logic_channel_open(struct euicc_ctx *ctx, const uint8_t *aid, uint8_t aid_len)
+static int apdu_interface_logic_channel_open(struct euicc_ctx *ctx, const uint8_t *aid, uint8_t aid_len)
 {
     char *response;
